+++ notion_df/entity/user.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from notion_df.utils.mixin import Dictable
+from notion_df.util.mixin import Dictable
 
 
 @dataclass
