+++ notion_df/form/request/create_database.py
@@ -1,19 +1,12 @@
 from __future__ import annotations
 
-from abc import abstractmethod, ABCMeta
+from abc import abstractmethod
 from dataclasses import dataclass
 
 from notion_df.form.common import RichText, Emoji, File
-from notion_df.utils.mixin import Dictable
-
-
-class RequestBuilder:
-    ENDPOINT = "https://api.notion.com/v1/"
-
-    @classmethod
-    @abstractmethod
-    def _get_entrypoint(cls):
-        pass
+from notion_df.form.request import RequestBuilder
+from notion_df.form.schema import PropertySchema
+from notion_df.util.mixin import Dictable
 
 
 class CreateDatabase(RequestBuilder):
@@ -23,12 +16,8 @@ class CreateDatabase(RequestBuilder):
         return "https://api.notion.com/v1/databases/"
 
 
-class RequestParams(Dictable, metaclass=ABCMeta):
-    ...
-
-
 @dataclass
-class CreateDatabaseParams(RequestParams):
+class CreateDatabaseParams(Dictable):
     parent_id: str
     icon: Emoji | File
     cover: File
@@ -46,7 +35,3 @@ class CreateDatabaseParams(RequestParams):
             "title": [rich_text.to_dict() for rich_text in self.title],
             "properties": {name: schema.to_dict() for name, schema in self.properties.items()}
         }
-
-
-class PropertySchema(Dictable, metaclass=ABCMeta):
-    ...
