+++ notion_df/form/schema.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta
+
+from notion_df.util.mixin import Dictable
+
+
+class PropertySchema(Dictable, metaclass=ABCMeta):
+    ...
