+++ notion_df/util/mixin.py
@@ -1,5 +1,7 @@
 from abc import ABC, abstractmethod
-from typing import Any
+from typing import Any, Generic, TypeVar
+
+_T = TypeVar('_T')
 
 
 class Dictable(ABC):
@@ -11,3 +13,7 @@ class Dictable(ABC):
 #     @classmethod
 #     @abstractmethod
 #     def from_dict(cls, value: dict) -> Self: ...
+class Promise(Generic[_T]):
+    @abstractmethod
+    def resolve(self) -> _T:
+        pass
